[YS-560] refactor: 유틸 함수 디렉토리 분리
